py/uniai/pyproject.toml
[project]
name = "uniai"
version = "0.0.1"
description = "An AI that helps you manage your technology!"
authors = [
    {name = "Dark Star☀🎵", email = "darkstar@unisocial.net"},
]
dependencies = [
    "python-vagrant>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "GPLv3"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
